+++ src/Nfts/NftsNavigator.tsx
@@ -1,5 +1,6 @@
 import {createStackNavigator} from '@react-navigation/stack'
 import React from 'react'
+import {defineMessages, useIntl} from 'react-intl'
 
 import {NftRoutes} from '../navigation'
 import {SearchProvider} from '../Search'
@@ -17,7 +18,8 @@ export const NftsNavigator = () => {
 }
 
 const Routes = () => {
-  const {searchHeaderOptions} = useSearchHeaderOptions()
+  const strings = useStrings()
+  const {searchHeaderOptions} = useSearchHeaderOptions({placeHolderText: strings.search, title: strings.title})
 
   return (
     <Stack.Navigator>
@@ -25,3 +27,23 @@ const Routes = () => {
     </Stack.Navigator>
   )
 }
+
+const useStrings = () => {
+  const intl = useIntl()
+
+  return {
+    title: intl.formatMessage(messages.title),
+    search: intl.formatMessage(messages.search),
+  }
+}
+
+const messages = defineMessages({
+  title: {
+    id: 'nft.navigation.title',
+    defaultMessage: '!!!NFT Gallery',
+  },
+  search: {
+    id: 'nft.navigation.search',
+    defaultMessage: '!!!Search NFT',
+  },
+})
